requirements.txt
asgiref==3.7.2
Django==5.0.1
djangorestframework==3.14.0
pillow==10.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
pytz==2024.1
sqlparse==0.4.4
